third_party/blink/web_tests/external/wpt/cookiestore/cookieStore_set_maxAge.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] cookieStore.set with maxAge set to a negative value
  assert_equals: expected null but got object "[object Object]"
[FAIL] cookieStore.set fails with both maxAge and expires
  assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
